Move vault token created handler to vault v2 namespace
